fix type not compatible to parsedqs

+++ Backend/controllers/transactionController.ts
@@ -1,23 +1,29 @@
 import express from "express";
 import * as transactionService from "../services/transactionsService.js";
+import type { Request } from "express";
 import type {
   GetUserControl,
   GetUserTransactionControl,
   NewTransactionControl,
-  SumTransactionsControl,
-  deleteTransactionControl,
+  DeleteTransactionControl,
 } from "../interfaces/interfaces.js";
 export const getUserTransaction = async (
-  req: GetUserTransactionControl,
+  req: Request,
   res: express.Response,
   next: express.NextFunction,
 ) => {
   try {
+    if (!req.user) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
+
+    const { start, end, order } =
+      req.query as unknown as GetUserTransactionControl;
     const userTransactions = await transactionService.getUserTransaction(
       Number(req.user.userId),
-      req.query.start,
-      req.query.end,
-      req.query.order,
+      start,
+      end,
+      order,
     );
     res.json(userTransactions);
   } catch (err) {
@@ -26,11 +32,14 @@ export const getUserTransaction = async (
 };
 
 export const newTransaction = async (
-  req: NewTransactionControl,
+  req: Request<{}, {}, NewTransactionControl>,
   res: express.Response,
   next: express.NextFunction,
 ) => {
   try {
+    if (!req.user) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
     const addTransactions = await transactionService.newTransaction({
       userId: req.user.userId,
       amount: req.body.amount,
@@ -45,14 +54,18 @@ export const newTransaction = async (
 };
 
 export const deleteTransaction = async (
-  req: deleteTransactionControl,
+  req: Request,
   res: express.Response,
   next: express.NextFunction,
 ) => {
   try {
+    if (!req.user) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
+    const { id } = req.params as unknown as DeleteTransactionControl;
     const deleteTransaction = await transactionService.deleteTransaction(
       req.user.userId,
-      req.params.id,
+      id,
     );
     res.json(deleteTransaction);
   } catch (err) {
@@ -61,11 +74,14 @@ export const deleteTransaction = async (
 };
 
 export const sumTransactions = async (
-  req: SumTransactionsControl,
+  req: Request<{}>,
   res: express.Response,
   next: express.NextFunction,
 ) => {
   try {
+    if (!req.user) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
     const sumTransactions = await transactionService.sumTransactions(
       req.user.userId,
     );
@@ -76,12 +92,16 @@ export const sumTransactions = async (
 };
 
 export const getUser = async (
-  req: GetUserControl,
+  req: Request,
   res: express.Response,
   next: express.NextFunction,
 ) => {
   try {
-    const user = await transactionService.getUser(req.query.userId);
+    if (!req.user) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
+    const { userId } = req.query as unknown as GetUserControl;
+    const user = await transactionService.getUser(userId);
     res.json(user);
   } catch (err) {
     return next(err);
